+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for popkey
 
+## 0.1.0.0
+- Drop profunctor instance in exchange for keyed folding
+- Add Store instance
+
 ## 0.0.0.1
 
 - Initial release